refactor(web)🛠️: Refactor web app structure and routing
- Extract database operations into a separate module for better separation of concerns.
- Create routers for logs and prompt versions to organize endpoint handling.
- Update HTML templates to reflect new routing paths.
