+++ llamabot/web/templates/log_table.html
@@ -12,7 +12,7 @@
     <tbody>
         {% if logs %}
             {% for log in logs %}
-            <tr class="log-row" data-prompt-hashes="{{ log.prompt_hashes }}" hx-get="/log/{{ log.id }}" hx-target="#log-details">
+            <tr class="log-row" data-prompt-hashes="{{ log.prompt_hashes }}" hx-get="/logs/{{ log.id }}" hx-target="#log-details">
                 <td>{{ log.id }}</td>
                 <td>{{ log.object_name }}</td>
                 <td>{{ log.timestamp }}</td>
