+++ llamabot/web/templates/message_log.html
@@ -12,7 +12,7 @@
                 <div class="prompt-info">
                     <span class="prompt-name">{{ message.prompt_name }}</span>
                     <a href="#" class="prompt-hash"
-                       hx-get="/prompt/{{ message.prompt_hash }}"
+                       hx-get="/prompts/{{ message.prompt_hash }}"
                        hx-target="#modal-content"
                        hx-trigger="click"
                        hx-swap="innerHTML">
