+++ llamabot/web/templates/rating_buttons.html
@@ -2,14 +2,14 @@
     <strong>Rate this response:</strong><br>
     <div class="btn-group mt-2" role="group">
         <button class="btn btn-outline-success me-2 rating-btn"
-                hx-post="/log/{{ log_id }}/rate"
+                hx-post="/logs/{{ log_id }}/rate"
                 hx-vals='{"rating": 1}'
                 hx-target="#rating-buttons-{{ log_id }}"
                 {% if rating == 1 %}disabled{% endif %}>
             👍 Helpful
         </button>
         <button class="btn btn-outline-danger rating-btn"
-                hx-post="/log/{{ log_id }}/rate"
+                hx-post="/logs/{{ log_id }}/rate"
                 hx-vals='{"rating": 0}'
                 hx-target="#rating-buttons-{{ log_id }}"
                 {% if rating == 0 %}disabled{% endif %}>
